+++ virtenc/enclave/globals.h
@@ -14,7 +14,7 @@ typedef struct _thread
     int sock;
     void* stack;
     size_t stack_size;
-    uint32_t tcs;
+    uint64_t tcs;
     int tid;
 } thread_t;
